# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LinkTokenResponse {
  linkToken: string;
  expiration: string;
}

export interface ExchangeTokenBody {
  publicToken: string;
  institutionName?: string;
}

export interface ExchangeTokenResponse {
  success: boolean;
  accountId: number;
}

export interface BankAccount {
  id: number;
  institutionName: string;
  accountName: string;
  accountType: string;
  /** @nullable */
  mask: string | null;
  connectedAt: string;
  transactionCount: number;
}

export interface SyncResult {
  added: number;
  modified: number;
  removed: number;
  total: number;
}

export interface TopMerchant {
  name: string;
  amount: number;
  visits: number;
  category: string;
}

export interface SpendingEra {
  id: number;
  name: string;
  tagline: string;
  season: string;
  year: number;
  startDate: string;
  endDate: string;
  totalSpent: number;
  topCategory: string;
  colorTheme: string;
  emoji: string;
  rank: number;
  description: string;
  categoryVibe?: string;
  topCategories: string[];
  topMerchants: TopMerchant[];
}

export interface CategoryBreakdown {
  category: string;
  amount: number;
  percentage: number;
  transactionCount: number;
}

export interface SpendingEraDetail {
  id: number;
  name: string;
  tagline: string;
  description: string;
  season: string;
  year: number;
  startDate: string;
  endDate: string;
  totalSpent: number;
  topCategory: string;
  colorTheme: string;
  emoji: string;
  rank: number;
  categoryVibe: string;
  categoryBreakdown: CategoryBreakdown[];
  topMerchants: TopMerchant[];
  weeklyAverage: number;
  peakWeek: string;
  funFact: string;
}

export interface ErasSummary {
  totalEras: number;
  totalSpent: number;
  biggestEra: string;
  mostFrequentCategory: string;
  dateRangeStart: string;
  dateRangeEnd: string;
  transactionCount: number;
  connected: boolean;
}