# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a Plaid Link token
 */
export const CreateLinkTokenResponse = zod.object({
  linkToken: zod.string(),
  expiration: zod.string(),
});

/**
 * @summary Exchange a public token for access token
 */
export const ExchangeTokenBody = zod.object({
  publicToken: zod.string(),
  institutionName: zod.string().optional(),
});

export const ExchangeTokenResponse = zod.object({
  success: zod.boolean(),
  accountId: zod.number(),
});

/**
 * @summary Get connected bank accounts
 */
export const GetAccountsResponseItem = zod.object({
  id: zod.number(),
  institutionName: zod.string(),
  accountName: zod.string(),
  accountType: zod.string(),
  mask: zod.string().nullable(),
  connectedAt: zod.string(),
  transactionCount: zod.number(),
});
export const GetAccountsResponse = zod.array(GetAccountsResponseItem);

/**
 * @summary Sync transactions from Plaid (2 years)
 */
export const SyncTransactionsResponse = zod.object({
  added: zod.number(),
  modified: zod.number(),
  removed: zod.number(),
  total: zod.number(),
});

/**
 * @summary Get all spending eras
 */
export const GetErasResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  tagline: zod.string(),
  season: zod.string(),
  year: zod.number(),
  startDate: zod.string(),
  endDate: zod.string(),
  totalSpent: zod.number(),
  topCategory: zod.string(),
  colorTheme: zod.string(),
  emoji: zod.string(),
  rank: zod.number(),
  description: zod.string(),
  categoryVibe: zod.string().optional(),
  topCategories: zod.array(zod.string()),
  topMerchants: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.number(),
      visits: zod.number(),
      category: zod.string(),
    }),
  ),
});
export const GetErasResponse = zod.array(GetErasResponseItem);

/**
 * @summary Analyze transactions and generate spending eras
 */
export const AnalyzeErasResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  tagline: zod.string(),
  season: zod.string(),
  year: zod.number(),
  startDate: zod.string(),
  endDate: zod.string(),
  totalSpent: zod.number(),
  topCategory: zod.string(),
  colorTheme: zod.string(),
  emoji: zod.string(),
  rank: zod.number(),
  description: zod.string(),
  categoryVibe: zod.string().optional(),
  topCategories: zod.array(zod.string()),
  topMerchants: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.number(),
      visits: zod.number(),
      category: zod.string(),
    }),
  ),
});
export const AnalyzeErasResponse = zod.array(AnalyzeErasResponseItem);

/**
 * @summary Get a single era with full details
 */
export const GetEraParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEraResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  tagline: zod.string(),
  description: zod.string(),
  season: zod.string(),
  year: zod.number(),
  startDate: zod.string(),
  endDate: zod.string(),
  totalSpent: zod.number(),
  topCategory: zod.string(),
  colorTheme: zod.string(),
  emoji: zod.string(),
  rank: zod.number(),
  categoryVibe: zod.string(),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      amount: zod.number(),
      percentage: zod.number(),
      transactionCount: zod.number(),
    }),
  ),
  topMerchants: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.number(),
      visits: zod.number(),
      category: zod.string(),
    }),
  ),
  weeklyAverage: zod.number(),
  peakWeek: zod.string(),
  funFact: zod.string(),
});

/**
 * @summary Get high-level summary stats across all eras
 */
export const GetErasSummaryResponse = zod.object({
  totalEras: zod.number(),
  totalSpent: zod.number(),
  biggestEra: zod.string(),
  mostFrequentCategory: zod.string(),
  dateRangeStart: zod.string(),
  dateRangeEnd: zod.string(),
  transactionCount: zod.number(),
  connected: zod.boolean(),
});