# lib/api-zod/src/generated/types/bankAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */

export interface BankAccount {
  id: number;
  institutionName: string;
  accountName: string;
  accountType: string;
  /** @nullable */
  mask: string | null;
  connectedAt: string;
  transactionCount: number;
}