# lib/api-zod/src/generated/types/erasSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */

export interface ErasSummary {
  totalEras: number;
  totalSpent: number;
  biggestEra: string;
  mostFrequentCategory: string;
  dateRangeStart: string;
  dateRangeEnd: string;
  transactionCount: number;
  connected: boolean;
}