# lib/api-zod/src/generated/types/linkTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */

export interface LinkTokenResponse {
  linkToken: string;
  expiration: string;
}