# lib/api-zod/src/generated/types/spendingEra.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */
import type { TopMerchant } from "./topMerchant";

export interface SpendingEra {
  id: number;
  name: string;
  tagline: string;
  season: string;
  year: number;
  startDate: string;
  endDate: string;
  totalSpent: number;
  topCategory: string;
  colorTheme: string;
  emoji: string;
  rank: number;
  description: string;
  categoryVibe?: string;
  topCategories: string[];
  topMerchants: TopMerchant[];
}