# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spending Eras API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncResult {
  added: number;
  modified: number;
  removed: number;
  total: number;
}